Migrate additional packages to .template.dependencies.json

+++ packages/react-scripts/scripts/init.js
@@ -115,16 +115,18 @@ module.exports = function(
   );
 
   let command;
-  let args;
+  let save;
+  let saveDev;
 
   if (useYarn) {
     command = 'yarnpkg';
-    args = ['add'];
+    save = ['add'];
   } else {
     command = 'npm';
-    args = ['install', '--save', verbose && '--verbose'].filter(e => e);
+    save = ['install', '--save', verbose && '--verbose'].filter(e => e);
+    saveDev = ['install', '--save-dev', verbose && '--verbose'].filter(e => e);
   }
-  args.push('react', 'react-dom');
+  save.push('react', 'react-dom');
 
   // Install additional template dependencies, if present
   const templateDependenciesPath = path.join(
@@ -133,11 +135,32 @@ module.exports = function(
   );
   if (fs.existsSync(templateDependenciesPath)) {
     const templateDependencies = require(templateDependenciesPath).dependencies;
-    args = args.concat(
+    save = save.concat(
       Object.keys(templateDependencies).map(key => {
         return `${key}@${templateDependencies[key]}`;
       })
     );
+    const templateDevDependencies = require(templateDependenciesPath).devDependencies;
+    saveDev = saveDev.concat(
+      Object.keys(templateDevDependencies).map(key => {
+        return `${key}@${templateDevDependencies[key]}`;
+      })
+    );
+
+    console.log(`Installing templated packages using ${command}...`);
+    console.log();
+
+    let proc = spawn.sync(command, save, { stdio: 'inherit' });
+    if (proc.status !== 0) {
+      console.error(`\`${command} ${save.join(' ')}\` failed`);
+      return;
+    }
+    proc = spawn.sync(command, saveDev, { stdio: 'inherit' });
+    if (proc.status !== 0) {
+      console.error(`\`${command} ${saveDev.join(' ')}\` failed`);
+      return;
+    }
+
     fs.unlinkSync(templateDependenciesPath);
   }
 
@@ -148,9 +171,14 @@ module.exports = function(
     console.log(`Installing react and react-dom using ${command}...`);
     console.log();
 
-    const proc = spawn.sync(command, args, { stdio: 'inherit' });
+    let proc = spawn.sync(command, save, { stdio: 'inherit' });
     if (proc.status !== 0) {
-      console.error(`\`${command} ${args.join(' ')}\` failed`);
+      console.error(`\`${command} ${save.join(' ')}\` failed`);
+      return;
+    }
+    proc = spawn.sync(command, saveDev, { stdio: 'inherit' });
+    if (proc.status !== 0) {
+      console.error(`\`${command} ${saveDev.join(' ')}\` failed`);
       return;
     }
   }
